+++ tests/mortar/tests.h
@@ -610,7 +610,7 @@ public:
         TrilinosWrappers::SparsityPattern dsp;
 
         dsp.reinit(dof_handler.locally_owned_dofs(),
-                   dof_handler.get_communicator());
+                   dof_handler.get_mpi_communicator());
 
         DoFTools::make_sparsity_pattern(dof_handler, dsp, *constraints);
 
@@ -1022,7 +1022,7 @@ public:
         TrilinosWrappers::SparsityPattern dsp;
 
         dsp.reinit(dof_handler.locally_owned_dofs(),
-                   dof_handler.get_communicator());
+                   dof_handler.get_mpi_communicator());
 
         DoFTools::make_sparsity_pattern(dof_handler, dsp, *constraints);
 
@@ -1390,7 +1390,7 @@ public:
         TrilinosWrappers::SparsityPattern dsp;
 
         dsp.reinit(dof_handler.locally_owned_dofs(),
-                   dof_handler.get_communicator());
+                   dof_handler.get_mpi_communicator());
 
         DoFTools::make_flux_sparsity_pattern(dof_handler, dsp, *constraints);
 
@@ -1823,7 +1823,7 @@ public:
         TrilinosWrappers::SparsityPattern dsp;
 
         dsp.reinit(dof_handler.locally_owned_dofs(),
-                   dof_handler.get_communicator());
+                   dof_handler.get_mpi_communicator());
 
         DoFTools::make_flux_sparsity_pattern(dof_handler, dsp, *constraints);